requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
numpy==1.24.3
pillow==10.2.0
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
requests==2.31.0
